+++ Helper/__init__.py
@@ -1,6 +1,5 @@
 from .log import Log, LogInfo
 from .log_level import Level
-from .config import Config, TapConfig
 from .child import Child
 from .serialize import Serialize
 from .tap_executor import Tap
